[python] split merge_into; align metrics and behavior with review
- Split data_evolution_merge_into.py into three modules:
  - data_evolution_merge_join.py (join construction + distributed write)
  - data_evolution_merge_transform.py (SET evaluation + transforms)
  - data_evolution_merge_into.py (entry, validation, scheduling)
- Return {num_matched, num_inserted, num_unchanged} per Jingsong review.
- Drop allow_multiple_matches; multi-match always errors.
- Silently skip blob columns during update (matches Spark).
- Defer global-index update action to a follow-up after PR #8045.
- Drop changes that belonged to PR #8045 (index_manifest_file.py,
  commit_message.py index_files field, file_store_commit.py
  index_deletes plumbing).
